+++ libraries/helpers.rb
@@ -2,10 +2,15 @@ module Gpg
   module Helpers
     include Chef::Mixin::ShellOut
 
-    def key_exists(new_resource)
+    def key_exists(new_resource, key = nil)
       gpg_check = gpg_cmd
-      gpg_check << gpg_opts if new_resource.override_default_keyring
-      gpg_check << "--list-keys | grep '#{new_resource.name_real}'"
+      gpg_check << override_command(new_resource) if new_resource.override_default_keyring
+
+      gpg_check << if new_resource.keyserver
+                     "--list-keys #{key}"
+                   else
+                     "--list-keys | grep #{new_resource.name_real}"
+                   end
 
       cmd = Mixlib::ShellOut.new(
         gpg_check,
@@ -14,30 +19,36 @@ module Gpg
       )
 
       cmd.run_command
-
       cmd.exitstatus == 0
     end
 
-    def gpg_opts(new_resource)
-      if new_resource.override_default_keyring
-        "--no-default-keyring --secret-keyring #{new_resource.secring_file} --keyring #{new_resource.pubring_file}"
-      else
-        false
-      end
+    def override_command(new_resource)
+      "--no-default-keyring --secret-keyring #{new_resource.secring_file} --keyring #{new_resource.pubring_file}"
     end
 
+    # Ensure GPG uses the correct home directory for the current resource
     def gpg_cmd
       "gpg2 --homedir #{new_resource.home_dir} "
     end
 
     def gpg2_packages
-      packages = %w(haveged)
-      if platform_family?('suse')
-        packages.push('gpg2')
+      case node['platform_family']
+      when 'suse'
+        %w(haveged gpg2)
+      when 'amazon'
+        if node['platform_version'].to_i >= 2023
+          %w(haveged)
+        else
+          %w(haveged gnupg2)
+        end
       else
-        packages.push('gnupg2')
+        %w(haveged gnupg2)
       end
-      packages
     end
   end
 end
+
+# package
+# gnupg2-minimal-2.3.7-1.amzn2023.0.4.aarch64 conflicts with
+# gnupg2 provided by
+# gnupg2-2.3.7-1.amzn2023.0.4.aarch64